chore(sonarlint): Fix sonarlint issues
The touched modules and classes have many sonarlint issues. This fixes a few of them.

+++ openrouteservice/src/main/java/org/heigit/ors/routing/pathprocessors/ExtraInfoProcessor.java
@@ -517,7 +517,7 @@ public class ExtraInfoProcessor implements PathProcessor {
 		if (osmIdInfoBuilder != null) {
 			long osmId = extOsmId.getEdgeValue(EdgeIteratorStateHelper.getOriginalEdge(edge));
 
-			osmIdInfoBuilder.addSegment((double)osmId, osmId, geom, dist);
+			osmIdInfoBuilder.addSegment(osmId, osmId, geom, dist);
 		}
 
 		if (roadAccessRestrictionsInfoBuilder != null) {
